# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP API",
        "description": "List of all api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employee/late/store": {
            "post": {
                "tags": [
                    "Employee Late Request"
                ],
                "summary": "Store employee late request",
                "operationId": "d0d0a4561c3592bea0a1dd5001951b96",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "arrival_time",
                                    "date",
                                    "reason"
                                ],
                                "properties": {
                                    "arrival_time": {
                                        "type": "string",
                                        "example": "12:35 PM"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2024-01-01"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Weeding Party"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully late request saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/late/show": {
            "get": {
                "tags": [
                    "Employee Late Request"
                ],
                "summary": "List employee late request",
                "operationId": "5e7610189e374c022aaa1603ce076ad3",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "numeric",
                                                        "example": "1"
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "example": "4"
                                                    },
                                                    "arrival_time": {
                                                        "type": "string",
                                                        "example": "10:20 AM"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2024-03-03"
                                                    },
                                                    "reason": {
                                                        "type": "string",
                                                        "example": "Weeding Party"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-03-14T11:03:57.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-03-14T11:03:57.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/leave/store": {
            "post": {
                "tags": [
                    "Employee Leave Request"
                ],
                "summary": "Store employee leave request",
                "operationId": "aad4d518052a88994b6f8349d0c82ed5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "leave_type_id",
                                    "from_date",
                                    "to_date",
                                    "reason"
                                ],
                                "properties": {
                                    "leave_type_id": {
                                        "type": "numeric",
                                        "example": "2"
                                    },
                                    "from_date": {
                                        "type": "string",
                                        "example": "2024-01-01"
                                    },
                                    "to_date": {
                                        "type": "string",
                                        "example": "2024-01-01"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Weeding Party"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully leave request saved."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employee/leave/show": {
            "get": {
                "tags": [
                    "Employee Leave Request"
                ],
                "summary": "List employee leave request",
                "operationId": "57dfbbf63c60c3895bb99209591b4034",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "sno": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "leave_type": {
                                                        "type": "string",
                                                        "example": "Annual"
                                                    },
                                                    "from_date": {
                                                        "type": "string",
                                                        "example": "2024-03-03"
                                                    },
                                                    "to_date": {
                                                        "type": "string",
                                                        "example": "2024-03-03"
                                                    },
                                                    "total_days": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "reason": {
                                                        "type": "string",
                                                        "example": "Weeding Party"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Pending"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}